% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_bfh_color_bar}
\alias{add_bfh_color_bar}
\title{Add BFH color bar annotation to plot}
\usage{
add_bfh_color_bar(
  plot,
  position = "top",
  color = bfh_cols("hospital_primary"),
  size = 0.02
)
}
\arguments{
\item{plot}{A ggplot2 object}

\item{position}{Position of the bar: "top", "bottom", "left", "right"}

\item{color}{Color of the bar. Default is BFH primary blue.}

\item{size}{Size of the bar (height for horizontal, width for vertical)}
}
\value{
A ggplot2 object with the color bar annotation
}
\description{
Adds a colored bar to the plot, useful for adding a visual BFH brand element.
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  theme_bfh()

add_bfh_color_bar(p, position = "top")
}
}
