% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{check_bfh_fonts}
\alias{check_bfh_fonts}
\title{Check if BFH fonts are installed}
\usage{
check_bfh_fonts()
}
\value{
Invisibly returns a named logical vector of font availability
}
\description{
Checks system for BFH preferred fonts and reports availability.
}
\examples{
\dontrun{
check_bfh_fonts()
}
}
