% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{check_colorblind_safe}
\alias{check_colorblind_safe}
\title{Check if colors are colorblind-friendly}
\usage{
check_colorblind_safe(colors)
}
\arguments{
\item{colors}{Character vector of hex colors}
}
\value{
Message about colorblind accessibility
}
\description{
Check if colors are colorblind-friendly
}
