% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_operators.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value to use if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the left-hand side if it's not NULL, otherwise returns the right-hand side.
This is similar to the \code{\%||\%} operator in rlang and purrr.
}
\examples{
\dontrun{
NULL \%||\% "default"      # "default"
"value" \%||\% "default"   # "value"
}
}
\keyword{internal}
