% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_bfh_continuous}
\alias{scale_fill_bfh_continuous}
\alias{scale_color_bfh_continuous}
\alias{scale_colour_bfh_continuous}
\title{Continuous colour scales using specific BFH palettes}
\usage{
scale_fill_bfh_continuous(palette = "blues", reverse = FALSE, ...)

scale_color_bfh_continuous(palette = "blues", reverse = FALSE, ...)

scale_colour_bfh_continuous(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette in [bfh_palettes()].}

\item{reverse}{Logical; reverse palette direction? Defaults to `FALSE`.}

\item{...}{Additional arguments passed to [ggplot2::scale_color_gradientn()].}
}
\description{
Continuous colour scales using specific BFH palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_bfh_continuous(palette = "blues")
}
}
\seealso{
[scale_color_bfh()], [bfh_pal()]

Other BFH scales: 
\code{\link{scale_bfh}},
\code{\link{scale_fill_bfh_discrete}()}
}
\concept{BFH scales}
