% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_bfh_discrete}
\alias{scale_fill_bfh_discrete}
\alias{scale_color_bfh_discrete}
\alias{scale_colour_bfh_discrete}
\title{Discrete color scales using specific BFH palettes}
\usage{
scale_fill_bfh_discrete(palette = "main", reverse = FALSE, ...)

scale_color_bfh_discrete(palette = "main", reverse = FALSE, ...)

scale_colour_bfh_discrete(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bfh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale()}
}
\description{
Discrete color scales using specific BFH palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_bfh_discrete(palette = "primary")
}
}
