% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{bfh_cols}
\alias{bfh_cols}
\title{Extract BFH colors as hex codes}
\usage{
bfh_cols(...)
}
\arguments{
\item{...}{Character names of bfh_colors}
}
\value{
Character vector of hex color codes
}
\description{
Extract BFH colors as hex codes
}
\examples{
# Get a single color
bfh_cols("primary")

# Get multiple colors
bfh_cols("hospital_primary", "hospital_blue")

# Get all colors
bfh_cols()

# Using aliases
bfh_cols("blue", "grey")
}
