% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{bfh_pal}
\alias{bfh_pal}
\title{Return function to interpolate a BFH color palette}
\usage{
bfh_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bfh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A function that takes an integer n and returns n colors
}
\description{
Return function to interpolate a BFH color palette
}
\examples{
bfh_pal("main")(5)
bfh_pal("blues", reverse = TRUE)(3)
}
