% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bfh_palettes}
\alias{bfh_palettes}
\title{BFH Color Palettes (named list)}
\format{
An object of class \code{list} of length 16.
}
\usage{
bfh_palettes
}
\description{
Predefined color palettes for different types of visualizations.
Includes palettes for both Hospital and Region Hovedstaden (koncern) branding.
}
\keyword{datasets}
