% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{bfh_palettes}
\alias{bfh_palettes}
\title{BFH Color Palettes (Named List)}
\format{
An object of class \code{list} of length 16.
}
\usage{
bfh_palettes
}
\description{
Predefined discrete and sequential palettes covering hospital, regional, and
legacy BFH use cases. Palettes are sourced from [bfh_colors].
}
\details{
The list is structured for use with [bfh_pal()] and the `scale_*_bfh*()`
helpers. Palettes ending in `"_seq"` are suited for continuous scales while
the rest target categorical data and infographics.
}
\seealso{
[bfh_pal()], [show_bfh_palettes()], [scale_color_bfh()]

Other BFH colors: 
\code{\link{bfh_colors}},
\code{\link{bfh_cols}()},
\code{\link{bfh_pal}()},
\code{\link{check_colorblind_safe}()},
\code{\link{clear_bfh_pal_cache}()},
\code{\link{show_bfh_palettes}()}
}
\concept{BFH colors}
\keyword{datasets}
