% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_bfh}
\alias{scale_bfh}
\alias{scale_color_bfh}
\alias{scale_colour_bfh}
\alias{scale_fill_bfh}
\title{BFH Color Scales for ggplot2}
\usage{
scale_color_bfh(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_colour_bfh(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_bfh(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bfh_palettes. Default is "main".}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A ggplot2 scale object
}
\description{
Color and fill scales using BFH color palettes for ggplot2.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_bfh()
}
\dontrun{
library(ggplot2)
ggplot(mtcars, aes(x = factor(cyl), fill = factor(cyl))) +
  geom_bar() +
  scale_fill_bfh()
}
}
