% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_bfh_continuous}
\alias{scale_fill_bfh_continuous}
\alias{scale_color_bfh_continuous}
\alias{scale_colour_bfh_continuous}
\title{Continuous color scales using specific BFH palettes}
\usage{
scale_fill_bfh_continuous(palette = "blues", reverse = FALSE, ...)

scale_color_bfh_continuous(palette = "blues", reverse = FALSE, ...)

scale_colour_bfh_continuous(palette = "blues", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in bfh_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to scale_color_gradientn()}
}
\description{
Continuous color scales using specific BFH palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_bfh_continuous(palette = "blues")
}
}
