% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_bfh_discrete}
\alias{scale_fill_bfh_discrete}
\alias{scale_color_bfh_discrete}
\alias{scale_colour_bfh_discrete}
\title{Discrete colour scales using specific BFH palettes}
\usage{
scale_fill_bfh_discrete(palette = "main", reverse = FALSE, ...)

scale_color_bfh_discrete(palette = "main", reverse = FALSE, ...)

scale_colour_bfh_discrete(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of a palette in [bfh_palettes()].}

\item{reverse}{Logical; reverse palette direction? Defaults to `FALSE`.}

\item{...}{Additional arguments passed to [ggplot2::discrete_scale()].}
}
\description{
Discrete colour scales using specific BFH palettes
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Sepal.Width, color = Species)) +
  geom_point() +
  scale_color_bfh_discrete(palette = "primary")
}
}
\seealso{
[scale_color_bfh()], [bfh_pal()], [bfh_palettes]

Other BFH scales: 
\code{\link{scale_bfh}},
\code{\link{scale_fill_bfh_continuous}()}
}
\concept{BFH scales}
