% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_bfh_palettes}
\alias{show_bfh_palettes}
\title{Show all available BFH color palettes}
\usage{
show_bfh_palettes(n = NULL)
}
\arguments{
\item{n}{Number of colors to display from each palette}
}
\description{
Show all available BFH color palettes
}
\examples{
\dontrun{
show_bfh_palettes()
}
}
